requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
pytest==8.0.2
fpdf2==2.7.8
python-dotenv==1.0.1
pytest-cov==4.1.0 